Added initial scripts
LoKi contains the basic functionality required to set up and solve equation 5. The test script contains a simple example of a single realisation of the model. ODEs,utils, and loaded_king are the three scripts i have been using during the development and are included here simply to show that the newest version produces the same psi(r) as the version which has been used to produce the figures in the manuscript.